requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-decouple
dj-database-url
gunicorn
psycopg2-binary